+++ src/conditions/FragmentEntryCondition.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace thepixelage\fragments\conditions;
+
+use craft\elements\conditions\entries\EntryCondition;
+
+class FragmentEntryCondition extends EntryCondition
+{
+    protected function conditionRuleTypes(): array
+    {
+        return array_merge(parent::conditionRuleTypes(), [
+            EntryUriConditionRule::class,
+        ]);
+    }
+}